requirements.txt
azure-functions
PyGithub 
requests